# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Headless CMS API",
    "version": "0.1.0",
    "description": "Production REST API for the Headless Content Management & Universal Template Render Engine."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local Development Server"
    },
    {
      "url": "https://cms.example.com/api/v1",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "HttpOnly session cookie created on Google OAuth login."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User JWT authorization header."
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Organization Delivery API Key (sk_live_...)."
      }
    },
...
  },
  "paths": {
...
    "/orgs/{orgId}/content/{slug}": {
      "get": {
        "summary": "List content entries",
        "tags": ["Content Entries"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "orgId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["DRAFT", "PUBLISHED"] } }
        ],
        "responses": {
          "200": { "description": "Paginated content entries" }
        }
      },
      "post": {
        "summary": "Create content entry",
        "tags": ["Content Entries"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "orgId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": { "type": "object" },
                  "publish": { "type": "boolean", "default": false }
                },
                "required": ["data"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Entry created" }
        }
      }
    },
...
    "/orgs/{orgId}/api-keys": {
      "get": {
        "summary": "List API keys",
        "tags": ["API Keys"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "orgId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of API keys" }
        }
      },
      "post": {
        "summary": "Generate new API key",
        "tags": ["API Keys"],
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [
          { "name": "orgId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Backend Service Key" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "API key generated with plaintext key value" }
        }
      }
    },
...
  }
}